extract TextualIO and completions into separate modules, fix startup blocking
- elroy/io/textual_io.py: TextualIO class (was inline in textual_app.py)
- elroy/io/completions.py: build_completions() and get_memory_panel_titles()
- _refresh_memory_panel and _update_completions promoted to @work(thread=True)
  so DB queries no longer block the Textual event loop on startup
