+++ tests/test_command.py
@@ -1,4 +1,3 @@
-
 from elroy.core.ctx import ElroyContext
 
 from .utils import process_test_message
